kubernetes: Inject step environment vars earlier
Many of the env vars added in PlanStep ae not available in Kubernetes
because the env is immutable after pod creation. So, to make these vars
available, we inject as many as possible in the CreateStep stage.

STEP_STATUS and STEP_STARTED don't make sense in CreateStep, however, so we
leave those out. They will just be empty for pipelines that get run by
the kubernetes runtime.